requirements.txt
hikari[speedups]>=2.0.0.dev111
hikari-lightbulb>=2.2.4
hikari-miru>=1.1.2
python-dotenv>=0.21.0